pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "privacyops-x"
version = "0.1.0"
description = "Safety-critical privacy operations benchmark for OpenEnv"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openenv-core[core]>=0.2.3",
    "fastapi>=0.135.0",
    "uvicorn>=0.43.0",
    "openai>=2.26.0",
    "httpx>=0.28.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
]

[project.scripts]
server = "privacyops_x.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["privacyops_x", "privacyops_x.server"]
package-dir = { "privacyops_x" = ".", "privacyops_x.server" = "server" }
